requirements.txt
pathway[all]
docling
python-dotenv
pandas
qrcode
pillow
litellm
google-generativeai